Adding keccakc (creativecoin) support
This is a very quick implementation of the keccakc variant for
creativecoin, which involves sha256d instead of sha256.

This patch is inspired by dan-and's patch in [1] and valided by mining
coins for the pool yiimp.eu, which worked at doubled rate compared to
cpuminer-multi (both compiled with -march=native).

[1] https://github.com/tpruvot/cpuminer-multi/pull/21/files